requirements.txt
flask
flask-cors
passlib[bcrypt]
pyjwt
google-auth
torch
numpy
pandas
langchain
langchain_community
langchain_core
scikit-learn
langchain_groq
chromadb
gunicorn
pypdf
sentence-transformers
langchain_text_splitters
ddgs
langgraph-checkpoint-sqlite